refactor: 멤버의 흐름으로 인한 중복코드 해결
참여자 리사이클러뷰에 관한 코드
