+++ app/src/main/java/com/stormers/storm/project/fragment/WaitingForStartingProjectFragment.kt
@@ -1,45 +1,20 @@
 package com.stormers.storm.project.fragment
 
-import android.content.Context
 import android.os.Bundle
-import android.util.Log
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
-import android.widget.Toast
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.stormers.storm.R
-import com.stormers.storm.base.BaseFragment
 import com.stormers.storm.customview.StormButton
 import com.stormers.storm.customview.dialog.StormDialog
 import com.stormers.storm.customview.dialog.StormDialogBuilder
 import com.stormers.storm.customview.dialog.StormDialogButton
-import com.stormers.storm.network.InterfaceProjectInfo
-import com.stormers.storm.network.InterfaceProjectUser
-import com.stormers.storm.network.RetrofitClient
 import com.stormers.storm.project.base.BaseProjectWaitingActivity
-import com.stormers.storm.project.model.ResponseProjectInfoModel
-import com.stormers.storm.project.model.ResponseProjectUserListModel
+import com.stormers.storm.round.base.BaseWaitingFragment
 import com.stormers.storm.round.fragment.HostRoundSettingFragment
-import com.stormers.storm.ui.HostRoundWaitingActivity
-import com.stormers.storm.user.ParticipantAdapter
-import com.stormers.storm.user.UserModel
-import com.stormers.storm.util.MarginDecoration
+import com.stormers.storm.round.fragment.MemberWaitingFragment
 import kotlinx.android.synthetic.main.activity_host_round_setting.*
 import kotlinx.android.synthetic.main.fragment_waiting_for_starting_project.*
-import kotlinx.android.synthetic.main.fragment_waiting_for_starting_project.view.*
-import kotlinx.android.synthetic.main.layout_list_of_participant.view.*
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
-class WaitingForStartingProjectFragment : BaseFragment(R.layout.fragment_waiting_for_starting_project) {
-
-    private val participantAdapter: ParticipantAdapter by lazy { ParticipantAdapter() }
+class WaitingForStartingProjectFragment : BaseWaitingFragment(R.layout.fragment_waiting_for_starting_project) {
 
     private lateinit var activityButton: StormButton
 
@@ -47,17 +22,9 @@ class WaitingForStartingProjectFragment : BaseFragment(R.layout.fragment_waiting
 
     private lateinit var dialog: StormDialog
 
-    private var projectIdx = -1
-
-    private lateinit var retrofitClient: InterfaceProjectUser
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
 
         super.onViewCreated(view, savedInstanceState)
-
-
-        projectIdx = (activity as BaseProjectWaitingActivity).projectIdx
-        showProjectUserList()
 
         activityButton = (activity as BaseProjectWaitingActivity).stormButton_ok_host_round_setting
 
@@ -69,17 +36,10 @@ class WaitingForStartingProjectFragment : BaseFragment(R.layout.fragment_waiting
                 if ((activity as BaseProjectWaitingActivity).isHost) {
                     goToFragment(HostRoundSettingFragment::class.java, null)
                 } else {
-                    //goToFragment(MemberRoundWaitingFragment::class.java, null)
+                    goToFragment(MemberWaitingFragment::class.java, null)
                 }
             }
         }
-
-        view.include_waitingproject_participant.recyclerview_participant.run {
-            layoutManager = LinearLayoutManager(context, RecyclerView.VERTICAL, false)
-            addItemDecoration(MarginDecoration(context, 15, RecyclerView.VERTICAL))
-            adapter = participantAdapter
-        }
-
 
         buttonArray.add(
             StormDialogButton("확인", false, object : StormDialogButton.OnClickListener {
@@ -99,28 +59,5 @@ class WaitingForStartingProjectFragment : BaseFragment(R.layout.fragment_waiting
         cardview_waitingproject_checkrules.setOnClickListener {
             fragmentManager?.let { it1 -> dialog.show(it1, "rule_reminder") }
         }
-    }
-
-    private fun showProjectUserList() {
-
-        retrofitClient = RetrofitClient.create(InterfaceProjectUser::class.java)
-
-        retrofitClient.getProjectUserList(projectIdx).enqueue(object : Callback<ResponseProjectUserListModel>{
-            override fun onFailure(call: Call<ResponseProjectUserListModel>, t: Throwable) {
-                Log.d("projectUser 통신실패","${t}")
-            }
-
-            override fun onResponse(
-                call: Call<ResponseProjectUserListModel>,
-                response: Response<ResponseProjectUserListModel>
-            ) {
-                if(response.isSuccessful)
-                    if(response.body()!!.success){
-                        Log.d("user list 통신성공","성공")
-                        participantAdapter.addAll(response.body()!!.data)
-                    }
-
-            }
-        })
     }
 }
