+++ app/src/main/java/com/stormers/storm/round/fragment/HostRoundSettingFragment.kt
@@ -1,12 +1,8 @@
 package com.stormers.storm.round.fragment
 
-import android.content.Intent
 import android.os.Bundle
 import android.util.Log
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.widget.Toast
 import com.stormers.storm.R
 import com.stormers.storm.base.BaseFragment
@@ -21,13 +17,11 @@ import com.stormers.storm.project.network.InterfaceRoundCount
 import com.stormers.storm.round.model.ResponseRoundCountModel
 import com.stormers.storm.round.model.RoundSettingModel
 import com.stormers.storm.ui.HostRoundWaitingActivity
-import com.stormers.storm.ui.RoundProgressActivity
 import kotlinx.android.synthetic.main.activity_host_round_setting.*
 import kotlinx.android.synthetic.main.fragment_host_round_setting.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.Retrofit
 import java.lang.StringBuilder
 
 class HostRoundSettingFragment : BaseFragment(R.layout.fragment_host_round_setting) {
