+++ app/src/main/java/com/stormers/storm/round/fragment/MemberWaitingFragment.kt
@@ -1,0 +1,15 @@
+package com.stormers.storm.round.fragment
+
+import android.os.Bundle
+import android.view.View
+import com.stormers.storm.R
+import com.stormers.storm.round.base.BaseWaitingFragment
+
+
+class MemberWaitingFragment : BaseWaitingFragment(R.layout.fragment_round_setting_waiting_member) {
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+    }
+}
