+++ app/src/main/java/com/stormers/storm/round/fragment/RoundStartFragment.kt
@@ -16,6 +16,7 @@ import com.stormers.storm.base.BaseFragment
 import com.stormers.storm.customview.StormButton
 import com.stormers.storm.customview.dialog.StormDialog
 import com.stormers.storm.customview.dialog.StormDialogBuilder
+import com.stormers.storm.round.base.BaseWaitingFragment
 import com.stormers.storm.ui.HostRoundWaitingActivity
 import com.stormers.storm.ui.RoundProgressActivity
 import com.stormers.storm.util.MarginDecoration
@@ -25,10 +26,7 @@ import kotlinx.android.synthetic.main.fragment_round_start.view.*
 import kotlinx.android.synthetic.main.layout_list_of_participant.view.*
 
 
-class RoundStartFragment : BaseFragment(R.layout.fragment_round_start) {
-
-    private val participantAdapter: ParticipantAdapter by lazy { ParticipantAdapter() }
-
+class RoundStartFragment : BaseWaitingFragment(R.layout.fragment_round_start) {
     private lateinit var activityButton: StormButton
 
     private lateinit var dialog: StormDialog
@@ -40,17 +38,6 @@ class RoundStartFragment : BaseFragment(R.layout.fragment_round_start) {
 
         initActivityButton()
 
-        initParticipant(view)
-    }
-
-    private fun initParticipant(view: View) {
-        view.include_roundstart_participant.recyclerview_participant.run {
-            layoutManager = LinearLayoutManager(context, RecyclerView.VERTICAL, false)
-            addItemDecoration(MarginDecoration(context, 15, RecyclerView.VERTICAL))
-            adapter = participantAdapter
-        }
-
-        participantAdapter.addAll(loadDatas())
     }
 
     private fun initActivityButton() {
@@ -66,30 +53,5 @@ class RoundStartFragment : BaseFragment(R.layout.fragment_round_start) {
 
             handler.postDelayed(handlerTask, 5000)
         }
-    }
-
-    fun loadDatas(): MutableList<UserModel> {
-        val datas = mutableListOf<UserModel>()
-
-        datas.add(
-            UserModel(
-                "https://www.notion.so/STORM-e0234061dd594af79f1035691830e698#8f611dc7d34b42f785d65cf7cc7a95bb",
-                "김성규"
-            )
-        )
-        datas.add(
-            UserModel(
-                "https://www.notion.so/STORM-e0234061dd594af79f1035691830e698#0a957fd1e94d43739b018f87d3cadd2b",
-                "손평화"
-            )
-        )
-        datas.add(
-            UserModel(
-                "https://www.notion.so/STORM-e0234061dd594af79f1035691830e698#56815b6b35c347109dc3bd3434bd6041",
-                "강희원"
-            )
-        )
-
-        return datas
     }
 }
