+++ app/src/main/res/layout/fragment_round_setting_waiting_member.xml
@@ -4,8 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/fragment_round_start"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".round.fragment.RoundStartFragment">
+    android:layout_height="match_parent">
 
 
     <FrameLayout
